# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "LIBRARY API documentation",
    "description": "Documentation of Library API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "contact@library-api.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8080",
  "servers": [
    {
      "url": "http://localhost:8080/"
    }
  ],
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "authorization",
      "scheme": "bearer",
      "bearerFormat": "JWT",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/libraryapi/api/v1/services/getbooks": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get all books in the library",
        "description": "This API endpoint retrieves all books in the library and returns a list of books in JSON format.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/libraryapi/api/v1/services/addbook": {
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Add a new book to the library",
        "description": "This API endpoint creates a new book in the library by sending a JSON payload in the request body.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Title": {
                  "example": "any"
                },
                "Author": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/updatebook/{bookId}": {
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "Update an existing book in the library",
        "description": "This API endpoint updates an existing book in the library by sending a JSON payload in the request body.",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/deletebook/{bookId}": {
      "delete": {
        "tags": [
          "Books"
        ],
        "summary": "Delete a book from the library",
        "description": "This API endpoint deletes a book from the library.",
        "parameters": [
          {
            "name": "bookId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/getauthors": {
      "get": {
        "tags": [
          "Authors"
        ],
        "summary": "Get authors",
        "description": "Get all authors from the database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/libraryapi/api/v1/services/addauthor": {
      "post": {
        "tags": [
          "Authors"
        ],
        "summary": "Add author",
        "description": "Add an author to the database",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Name": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/updateauthor/{authorId}": {
      "put": {
        "tags": [
          "Authors"
        ],
        "summary": "Update author",
        "description": "Update an author in the database",
        "parameters": [
          {
            "name": "authorId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/deleteauthor/{authorId}": {
      "delete": {
        "tags": [
          "Authors"
        ],
        "summary": "Delete author",
        "description": "Delete an author from the database",
        "parameters": [
          {
            "name": "authorId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/getgenres": {
      "get": {
        "tags": [
          "Genres"
        ],
        "summary": "Get all genres",
        "description": "This endpoint is used to retrieve all genres from the database.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/libraryapi/api/v1/services/addgenre": {
      "post": {
        "tags": [
          "Genres"
        ],
        "summary": "Add a new genre",
        "description": "This endpoint is used to add a new genre to the database.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Name": {
                  "example": "any"
                },
                "Description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/updategenre/{genreId}": {
      "put": {
        "tags": [
          "Genres"
        ],
        "summary": "Update a genre",
        "description": "This endpoint is used to update an existing genre in the database.",
        "parameters": [
          {
            "name": "genreId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/deletegenre/{genreId}": {
      "delete": {
        "tags": [
          "Genres"
        ],
        "summary": "Delete a genre",
        "description": "This endpoint is used to delete a genre from the database.",
        "parameters": [
          {
            "name": "genreId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/getloans": {
      "get": {
        "tags": [
          "Loans"
        ],
        "summary": "Get all loan information",
        "description": "Retrieve all loan information from the database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/libraryapi/api/v1/services/addloan": {
      "post": {
        "tags": [
          "Loans"
        ],
        "summary": "Add a new loan",
        "description": "Add a new loan to the database",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "BookName": {
                  "example": "any"
                },
                "DueDate": {
                  "example": "any"
                },
                "LoanDate": {
                  "example": "any"
                },
                "MemberName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/updateloan/{loanId}": {
      "put": {
        "tags": [
          "Loans"
        ],
        "summary": "Update a loan",
        "description": "Update an existing loan in the database",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/libraryapi/api/v1/services/deleteloan/{loanId}": {
      "delete": {
        "tags": [
          "Loans"
        ],
        "summary": "Delete a loan",
        "description": "Delete an existing loan from the database",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}